Consistency in integer overflow checks
