+++ packages/tx/test/base.spec.ts
@@ -9,7 +9,7 @@ import {
 } from '../src'
 import { TxsJsonEntry } from './types'
 import { BaseTransaction } from '../src/baseTransaction'
-import { privateToPublic, BN, toBuffer, MAX_INTEGER, MAX_UINT64 } from 'ethereumjs-util'
+import { privateToPublic, BN, toBuffer, MAX_INTEGER } from 'ethereumjs-util'
 
 tape('[BaseTransaction]', function (t) {
   // EIP-2930 is not enabled in Common by default (2021-03-06)
@@ -400,28 +400,6 @@ tape('[BaseTransaction]', function (t) {
       st.ok(
         err.message.includes('unimplemented bits value'),
         'throws when bits value other than 64 or 256 provided'
-      )
-    }
-
-    const validAddress = Buffer.from('01'.repeat(20), 'hex')
-    const validSlot = Buffer.from('01'.repeat(32), 'hex')
-    const chainId = new BN(1)
-    try {
-      AccessListEIP2930Transaction.fromTxData(
-        {
-          data: Buffer.from('010200', 'hex'),
-          to: validAddress,
-          accessList: [[validAddress, [validSlot]]],
-          chainId,
-          gasLimit: MAX_UINT64,
-          gasPrice: MAX_INTEGER,
-        },
-        { common }
-      )
-    } catch (err: any) {
-      st.ok(
-        err.message.includes('gasPriceXgasLimit cannot exceed MAX_INTEGER'),
-        'throws when gasLimit * gasPrice exceeds MAX_INTEGER'
       )
     }
     st.end()
